Using implicit DLL linking
Fixed bug with deleting the camera in another DLL than it was allocated
in.
Fixed bug when loading images
Fixed bug with the current display mode in the renderer
Cleaned up the renderer
Simplified one of the rendering methods
Output directories in Windows are now the same as on Linux, ignoring the
current build config and being outputed to the bin folder.